requirements.txt
icecream==2.1.1
numpy==1.20.3
face_recognition==1.3.0
numpy_serializer==0.1.3
Pillow==8.3.1
SQLAlchemy==1.4.21
